requirements.txt
pyspark
apache-airflow-providers-postgres
psycopg2-binary
apache-airflow-providers-apache-spark